elearning_app/requirements/base.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
django-webpack-loader==0.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
django-environ==0.4.5
environ~=1.0